+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - [Policy] Fix #3518 rewrite de-indexing policy (#3993 by: JonnyOThan; reviewed: HebaruSan)
 - [Netkan] Fix null reference exception in swinfo transformer (#3999 by: HebaruSan)
 - [Netkan] Improve netkan relationship error message (#4020, #4021 by: HebaruSan)
+- [Core] Get KSP2 version from game assembly (#4034 by: HebaruSan)
 
 ## v1.34.4 (Niven)